incoporate mantine theme + dark/light mode to storybook interface
