+++ client/package.json
@@ -37,6 +37,7 @@
     "eslint-plugin-storybook": "^0.6.15",
     "prop-types": "^15.8.1",
     "storybook": "^7.6.10",
+    "storybook-dark-mode": "^3.0.3",
     "vite": "^5.0.8"
   }
 }
