+++ client/src/theme.js
@@ -1,0 +1,7 @@
+import {createTheme} from "@mantine/core"
+
+export const theme = createTheme({
+//themes here 
+  fontFamily: 'serif'
+
+})
